chore: move scheuler messages from pkg/scheduler to internal/sproto
