+++ master/internal/agent/slots.go
@@ -7,11 +7,11 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/determined-ai/determined/master/internal/api"
+	"github.com/determined-ai/determined/master/internal/sproto"
 	"github.com/determined-ai/determined/master/pkg/actor"
 	aproto "github.com/determined-ai/determined/master/pkg/agent"
 	"github.com/determined-ai/determined/master/pkg/check"
 	"github.com/determined-ai/determined/master/pkg/container"
-	"github.com/determined-ai/determined/master/pkg/scheduler"
 )
 
 type slots struct {
@@ -40,7 +40,7 @@ func (s *slots) Receive(ctx *actor.Context) error {
 		for _, child := range ctx.Children() {
 			ctx.Tell(child, msg)
 		}
-	case scheduler.ContainerStateChanged:
+	case sproto.ContainerStateChanged:
 		s.sendToSlots(ctx, msg.Container, msg)
 	case echo.Context:
 		s.handleAPIRequest(ctx, msg)
