+++ master/internal/sproto/resource_providers.go
@@ -1,4 +1,4 @@
-package scheduler
+package sproto
 
 import (
 	"bytes"
